week_06/transformer-app/app/requirements.txt
attrdict==2.0.1
pymongo==3.12.1
python-dotenv==0.19.1
schedule==1.1.0
SQLAlchemy
pandas==1.3.3
psycopg2-binary==2.9.1